+++ testing/sct_dmri_moco/test_sct_dmri_moco.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 #
-# This script tests sct_warp_atlas2metric.
-#
 
 # subject list
 SUBJECT_LIST="errsm_23" 
@@ -37,11 +35,12 @@ for subject in $SUBJECT_LIST; do
         -i ../../data/${subject}/${contrast}/${file1}
         -b ../../data/${subject}/${contrast}/${file2}
         -v 1
+        -z 0
         -f 0
         -s 15
         -d 3
-        -r 0
-        -p sinc"
+        -r 1
+        -p spline"
     echo ==============================================================================================
     echo "$cmd"
     echo ==============================================================================================